Refactor entrypoint
